Moved the old gui to an new package and started a new one in code
