+++ JAICore/jaicore-graphvisualizer/src/jaicore/graphvisualizer/guiOld/FXGui.java
