+++ JAICore/jaicore-graphvisualizer/src/jaicore/graphvisualizer/guiOld/GuiApplication.java
