+++ JAICore/jaicore-graphvisualizer/src/jaicore/graphvisualizer/guiOld/ReconstructionNodeDataSupplier.java
