+++ JAICore/jaicore-graphvisualizer/src/jaicore/graphvisualizer/guiOld/Recorder.java
