+++ JAICore/jaicore-graphvisualizer/src/jaicore/graphvisualizer/guiOld/dataSupplier/ISupplier.java
